# docu/rest/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.2",
        "title": "Chimera API"
    },
    "host": "localhost:8080",
    "basePath": "/api/interface/v2",
    "schemes": [
        "http"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
        "/scenario": {
            "get": {
                "description": "This method gives an overview of all current scenarios.",
                "operationId": "getScenarios_",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "filter",
                        "in": "query",
                        "required": false,
                        "x-is-map": false,
                        "type": "string",
                        "description": "Specifies a search. Only scenarios which\n                     name contain the specified string will be\n                     returned."
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of all scenarios.",
                        "schema": {
                            "$ref": "#/definitions/ScenarioList"
                        }
                    }
                }
            },
            "post": {
                "description": "This method allows to upload a new scenario.",
                "operationId": "postInstance_",
                "consumes": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "scenario",
                        "in": "body",
                        "required": true,
                        "x-is-map": false,
                        "schema": {
                            "$ref": "#/definitions/Scenario"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successfully created scenario."
                    },
                    "400": {
                        "description": "Invalid fragment XML.",
                        "schema": {
                            "$ref": "#/definitions/ErrorArray"
                        }
                    },
                    "422": {
                        "description": "Failed to process the scenario.",
                        "schema": {
                            "$ref": "#/definitions/ErrorArray"
                        }
                    }
                }
            }
        },
        "/scenario/{scenarioID}": {
            "get": {
                "description": "This method provides information about one scenario.",
                "operationId": "getScenario_",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "scenarioID",
                        "in": "path",
                        "required": true,
                        "x-is-map": false,
                        "type": "integer",
                        "format": "int64",
                        "description": "The Id of the scenario used inside the database."
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detailed information about the scenario.",
                        "schema": {
                            "$ref": "#/definitions/ScenarioDetails"
                        }
                    },
                    "404": {
                        "description": "The scenario wasn't found.",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
...
        "/scenario/{scenarioID}/instance": {
            "get": {
                "description": "This method provides information about all instances of one scenario.",
                "operationId": "getScenarioInstances_",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "scenarioID",
                        "in": "path",
                        "required": true,
                        "x-is-map": false,
                        "type": "integer",
                        "format": "int64",
                        "description": "The id of the scenario which instances should be returned."
                    },
                    {
                        "name": "filter",
                        "in": "query",
                        "required": false,
                        "x-is-map": false,
                        "type": "string",
                        "description": "Specifies a search. Only scenarios which\n                     name contain the specified string will be\n                     returned."
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of all instances for the scenario.",
                        "schema": {
                            "$ref": "#/definitions/ScenarioInstanceList"
                        }
                    },
                    "404": {
                        "description": "The scenario wasn't found.",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            },
            "put": {
                "description": "Creates a new instance of a specified scenario. The name of the instance can be specified in the request.",
                "operationId": "Update_startNewNamedInstance_",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "scenarioID",
                        "in": "path",
                        "required": true,
                        "x-is-map": false,
                        "type": "integer",
                        "format": "int64",
                        "description": "the id of the scenario."
                    },
                    {
                        "name": "name",
                        "in": "body",
                        "required": true,
                        "x-is-map": false,
                        "description": "The name, which will be used for the new instance.",
                        "schema": {
                            "$ref": "#/definitions/NamedJaxBean"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "successfully created the instance.",
                        "schema": {
                            "$ref": "#/definitions/ScenarioRef"
                        }
                    },
                    "404": {
                        "description": "The scenario wasn't found.",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a new instance of a specified scenario. The instance will have the same name as the scenario.",
                "operationId": "Create_startNewInstance_",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "scenarioID",
                        "in": "path",
                        "required": true,
                        "x-is-map": false,
                        "type": "integer",
                        "format": "int64",
                        "description": "the id of the scenario."
                    }
                ],
                "responses": {
                    "201": {
                        "description": "successfully created the instance.",
                        "schema": {
                            "$ref": "#/definitions/ScenarioRef"
                        }
                    },
                    "404": {
                        "description": "The scenario wasn't found.",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
        "/scenario/{scenarioID}/instance/{instanceID}": {
            "get": {
                "description": "This method provides detailed information about a scenario instance.",
                "operationId": "getScenarioInstance_",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "scenarioID",
                        "in": "path",
                        "required": true,
                        "x-is-map": false,
                        "type": "integer",
                        "format": "int64",
                        "description": "The ID of the scenario."
                    },
                    {
                        "name": "instanceID",
                        "in": "path",
                        "required": true,
                        "x-is-map": false,
                        "type": "integer",
                        "format": "int64",
                        "description": "The ID of the instance."
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detailed information about the instance.",
                        "schema": {
                            "$ref": "#/definitions/ScenarioInstance"
                        }
                    },
                    "404": {
                        "description": "The scenario or instance wasn't found.",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
        "/scenario/{scenarioID}/instance/{instanceID}/events": {
            "get": {
                "description": "This method provides a list of registered events for the instance.",
                "operationId": "getEvents_",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "scenarioID",
                        "in": "path",
                        "required": true,
                        "x-is-map": false,
                        "type": "integer",
                        "format": "int64",
                        "description": "The ID of the scenario."
                    },
                    {
                        "name": "instanceID",
                        "in": "path",
                        "required": true,
                        "x-is-map": false,
                        "type": "integer",
                        "format": "int64",
                        "description": "The ID of the instance."
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of events for the scenario instance.",
                        "schema": {
                            "$ref": "#/definitions/EventList"
                        }
                    },
                    "404": {
                        "description": "The scenario or instance wasn't found.",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
        "/scenario/{scenarioID}/instance/{instanceID}/canTerminate": {
            "get": {
                "description": "This determines whether a given scenario instance can terminate.",
                "operationId": "checkTermination_",
                "produces": [
                    "text/plain",
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "scenarioID",
                        "in": "path",
                        "required": true,
                        "x-is-map": false,
                        "type": "integer",
                        "format": "int64",
                        "description": "The ID of the scenario."
                    },
                    {
                        "name": "instanceID",
                        "in": "path",
                        "required": true,
                        "x-is-map": false,
                        "type": "integer",
                        "format": "int64",
                        "description": "The ID of the instance."
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Instance can be deleted",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Termination condition is not fulfilled",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "The scenario or instance wasn't found.",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
        "/scenario/{scenarioID}/instance/{instanceID}/terminate": {
            "post": {
                "description": "This method is used to terminate a scenario instance.",
                "operationId": "Create_terminateScenarioInstance_",
                "produces": [
                    "text/plain",
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "scenarioID",
                        "in": "path",
                        "required": true,
                        "x-is-map": false,
                        "type": "integer",
                        "format": "int64",
                        "description": "The ID of the scenario."
                    },
                    {
                        "name": "instanceID",
                        "in": "path",
                        "required": true,
                        "x-is-map": false,
                        "type": "integer",
                        "format": "int64",
                        "description": "The ID of the instance."
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Instance has been terminated",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Termination condition is not fulfilled",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "The scenario or instance wasn't found.",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
        "/scenario/{scenarioID}/instance/{instanceID}/activity": {
            "get": {
                "description": "Returns a JSON-Object containing information about all activity\n instances of a specified scenario instance.",
                "operationId": "getActivitiesOfInstance_",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "scenarioID",
                        "in": "path",
                        "required": true,
                        "x-is-map": false,
                        "type": "integer",
                        "format": "int64",
                        "description": "The id of the scenario"
                    },
                    {
                        "name": "instanceID",
                        "in": "path",
                        "required": true,
                        "x-is-map": false,
                        "type": "integer",
                        "format": "int64",
                        "description": "The id of the instance."
                    },
                    {
                        "name": "filter",
                        "in": "query",
                        "required": false,
                        "x-is-map": false,
                        "type": "string",
                        "description": "Defines a search strings. Only activities\n                     with a label containing this String will be\n                     shown."
                    },
                    {
                        "name": "state",
                        "in": "query",
                        "required": false,
                        "x-is-map": false,
                        "type": "string",
                        "description": "The state of the instance."
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A list with all activity instances for the scenario instance.",
                        "schema": {
                            "$ref": "#/definitions/Activities"
                        }
                    },
                    "404": {
                        "description": "The scenario or instance wasn't found.",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "NamedJaxBean": {
            "title": "NamedJaxBean",
            "type": "object",
            "properties": {
                "name": {
                    "description": "The name which should be assigned to the entity.",
                    "type": "string"
                }
            }
        },
...
        "ScenarioList": {
            "type": "object",
            "properties": {
                "ids": {
                    "type": "array",
                    "description": "IDs of all known scenarios.",
                    "items": {
                        "type": "number",
                        "format": "integer"
                    }
                },
                "links": {
                    "type": "array",
                    "description": "References to all known scenarios.",
                    "items": {
                        "$ref": "#/definitions/ScenarioRef"
                    }
                },
                "labels": {
                    "type": "array",
                    "description": "Labels of all known scenarios.",
                    "items": {
                        "$ref": "#/definitions/ScenarioLabel"
                    }
                }
            }
        },
        "ScenarioRef": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "number",
                    "format": "integer",
                    "description": "ID of the scenario."
                },
                "link": {
                    "type": "string",
                    "description": "URI of the scenario."
                }
            }
        },
        "ScenarioLabel": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "number",
                    "format": "integer",
                    "description": "ID of the scenario."
                },
                "label": {
                    "type": "string",
                    "description": "Label of the scenario."
                }
            }
        },
        "ScenarioDetails": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "number",
                    "format": "integer",
                    "description": "The database ID of the scenario."
                },
                "name": {
                    "type": "string",
                    "description": "The name of the scenario."
                },
                "modelid": {
                    "type": "number",
                    "format": "integer",
                    "description": "The model ID of the scenario."
                },
                "modelversion": {
                    "type": "number",
                    "format": "integer",
                    "description": "The model version of the scenario."
                },
                "instances": {
                    "type": "string",
                    "description": "URI to retrieve instances of the scenario."
                }
            }
        },
        "ScenarioInstance": {
            "type": "object",
            "properties": {
                "activities": {
                    "type": "string",
                    "description": "URI to retrieve activities in the scenarioInstance."
                },
                "name": {
                    "type": "string",
                    "description": "The name of the scenarioInstance."
                },
                "id": {
                    "type": "number",
                    "format": "integer",
                    "description": "The ID of the scenarioInstance."
                },
                "terminated": {
                    "type": "boolean"
                },
                "scenario_id": {
                    "type": "number",
                    "format": "integer",
                    "description": "The modelId of the scenario."
                }
            }
        },
...
        "ScenarioInstanceList": {
            "type": "object",
            "properties": {
                "ids": {
                    "type": "array",
                    "description": "IDs of all known scenario instances.",
                    "items": {
                        "type": "number",
                        "format": "integer"
                    }
                },
                "labels": {
                    "$ref": "#/definitions/InstanceLabel"
                },
                "links": {
                    "$ref": "#/definitions/InstanceRef"
                }
            }
        },
        "InstanceLabel": {
            "type": "object",
            "properties": {
                "(id)": {
                    "type": "string",
                    "description": "Instance id with the respective label."
                }
            }
        },
...
        "TerminationCondition": {
            "type": "object",
            "properties": {
                "conditions": {
                    "$ref": "#/definitions/TerminationConditionList"
                },
                "setIDs": {
                    "type": "array",
                    "description": "IDs of the termination conditions.",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
...
        "Scenario": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string",
                    "description": "The scenarios name"
                },
                "_id": {
                    "type": "number",
                    "format": "integer",
                    "description": "Unique identifier of any scenario."
                },
                "terminationconditions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/TerminationCondition"
                    },
                    "description": "The termination conditions that are used."
                },
                "domain_model": {
                    "type": "number",
                    "format": "integer",
                    "description": "The ID of the associated domain model."
                },
                "fragments": {
                    "type": "array",
                    "items": {
                        "type": "number",
                        "format": "integer"
                    },
                    "description": "The IDs of associated fragments."
                },
                "revision": {
                    "type": "number",
                    "format": "integer",
                    "description": "The scenarios revision. Is increased automatically if the\nfragments, the data_objects or the termination condition is different \nfrom the last version\n"
                }
            }
        },
...
        "EventList": {
            "type": "array",
            "description": "List of events.",
            "items": {
                "type": "string"
            }
        },
...
        "Activities": {
            "type": "object",
            "properties": {
                "activities": {
                    "$ref": "#/definitions/ActivityList"
                },
                "ids": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    },
                    "description": "List of all activity ids"
                }
            }
        },
...
        "Error": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "description": "Description of the error."
                }
            }
        },
        "ErrorArray": {
            "type": "array",
            "description": "List of errors",
            "items": {
                "$ref": "#/definitions/ErrorObject"
            }
        },
...
    }
}